[Data] Rename external shuffle operator files
Match the external_shuffle_* naming of tasks/runtime.
