+++ python/ray/data/_internal/logical/rules/operator_fusion.py
@@ -27,7 +27,7 @@ from ray.data._internal.execution.operators.map_operator import MapOperator
 from ray.data._internal.execution.operators.shuffle_operators.shuffle_reduce_operator import (
     ShuffleReduceOp,
 )
-from ray.data._internal.execution.operators.shuffle_operators.shuffle_reduce_operator_external import (  # noqa: E501
+from ray.data._internal.execution.operators.shuffle_operators.external_shuffle_reduce_operator import (  # noqa: E501
     ExternalHashShuffleReduceOp,
 )
 from ray.data._internal.execution.operators.task_pool_map_operator import (
