+++ python/ray/data/tests/test_hash_shuffle_external_smoke.py
@@ -18,10 +18,10 @@ from ray.data._internal.execution.operators.hash_shuffle_v2 import (
     _concat_reduce,
     _make_hash_partition_fn,
 )
-from ray.data._internal.execution.operators.shuffle_operators.shuffle_map_operator_external import (  # noqa: E501
+from ray.data._internal.execution.operators.shuffle_operators.external_shuffle_map_operator import (  # noqa: E501
     ExternalHashShuffleMapOp,
 )
-from ray.data._internal.execution.operators.shuffle_operators.shuffle_reduce_operator_external import (  # noqa: E501
+from ray.data._internal.execution.operators.shuffle_operators.external_shuffle_reduce_operator import (  # noqa: E501
     ExternalHashShuffleReduceOp,
 )
 from ray.data._internal.execution.util import make_ref_bundles
